chore: Few refactors
