+++ Enterwell.CI.Changelog.VSCodeExtension/src/constants.ts
@@ -2,3 +2,13 @@
  * Allowed types of changes. Types following the 'Keep a Changelog' guiding principles.
  */
 export const allowedTypes: string[] = ['Added', 'Changed', 'Deprecated', 'Removed', 'Fixed', 'Security'];
+
+/**
+ * Changelog configuration name.
+ */
+export const configurationName: string = '.changelog.json';
+
+/**
+ * Changelog changes directory name.
+ */
+export const changesDirectoryName: string = 'changes';
